requirements.txt
langchain
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
rank_bm25
gradio
python-dotenv
pandas
jq